feat: createIndex onPostBuild
